requirements.txt
absl-py==0.14.1
astunparse==1.6.3
blis==0.4.1
cached-property==1.5.2
cachetools==4.2.4
cairocffi==1.3.0
CairoSVG==2.5.2
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
colorama==0.4.4
cssselect2==0.4.1
cycler==0.10.0
cymem==2.0.5
defusedxml==0.7.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
Jinja2==3.0.2
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mplcursors==0.4
murmurhash==1.0.5
nltk==3.6.5
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pathy==0.6.0
Pillow==8.4.0
plac==1.1.3
preshed==3.0.5
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0
pycparser==2.20
pydantic==1.8.2
PyMuPDF==1.18.10
pyparsing==2.4.7
pyphen==0.11.0
python-dateutil==2.8.2
python-doctr==0.4.0
rapidfuzz==1.7.1
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
Shapely==1.7.1
six==1.15.0
smart-open==5.2.1
spacy==2.2.4
spacy-legacy==3.0.8
srsly==1.0.5
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
thinc==7.4.0
tinycss2==1.1.0
tqdm==4.62.3
typeguard==2.13.0
typer==0.4.0
typing-extensions==3.7.4.3
Unidecode==1.3.2
urllib3==1.26.7
wasabi==0.8.2
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
